refactor: extract startProxy into a separate file
